[back] fix: Todo delete 기능 완성

+++ instudy/src/main/java/instudy/instudy/service/TodoService.java
@@ -1,5 +1,6 @@
 package instudy.instudy.service;
 
+import instudy.instudy.domain.StudyStatus;
 import instudy.instudy.domain.Todo;
 import instudy.instudy.repository.TodoRepository;
 import org.springframework.stereotype.Service;
@@ -16,6 +17,7 @@ public class TodoService {
     public TodoService(TodoRepository todoRepository) {
         this.todoRepository = todoRepository;
     }
+
     public boolean join(Todo newTodo) {
         todoRepository.save(newTodo);
         return true;
@@ -25,8 +27,11 @@ public class TodoService {
         return todoRepository.findAll();
     }
 
-    public void deleteTodo(Todo todo) {
-        todoRepository.delete(todo);
+    public void deleteTodoByTodoText(String todoText) {
+        Todo todo = todoRepository.findByTodoText(todoText);
+        if (todo != null) {
+            todoRepository.delete(todo);
+        }
     }
 
     public void updateStatus(Todo todo) {
